moved reference, amdloader, html2ts, templatecache into their own files
